Update mk3PartSDCard.sh script

+++ mk3PartSDCard.sh
@@ -2,8 +2,14 @@
 # mk3PartSDCard.sh v0.3
 # Licensed under terms of GPLv2
 
-DRIVE=$1
+if [ -z $1 ]
+then
+	echo "mk3PartSDCard.sh Usage:"
+	echo "mk3PartSDCard.sh <device>"
+	exit
+fi
 
+DRIVE=$1
 dd if=/dev/zero of=$DRIVE bs=1024 count=1024
 
 SIZE=`fdisk -l $DRIVE | grep Disk | awk '{print $5}'`
@@ -21,4 +27,3 @@ mkfs.vfat -F 32 -n "boot" ${DRIVE}p1 > /dev/null 2>&1
 umount ${DRIVE}p1
 mkfs.ext4 -L "rootfs" ${DRIVE}p2 > /dev/null 2>&1
 umount ${DRIVE}p2
-#mkfs.ext4 -L "START_HERE" ${DRIVE}3
